refactor: BannerListWidget 컴포넌트 분리
