+++ apps/admin/src/features/banner/model/types.ts
@@ -1,0 +1,1 @@
+export type BannerFilterType = 'all' | 'active' | 'inactive';
